* trunk: (132 commits)
  DIC migration; logger migration
  DIC migration; logger migration;
  DIC migration; logging migration
  DIC migration, logger migration;
  fixed #22966 in removing prompt, no idea why this wass there in the first place
  fixed mantis 0023072: Course import failed : cannot import image without content (LoK)
  fixed missing disabled for manually selected partiticpants checkbox in test settings
  fixed mantis 0023040: Negative Brüche in Formelfragen weg radiert.
  User: Removed dead code
  User: Fixed #23092 (Administrators currently not forced to accept new user agreement)
  possible bug fix for mantis ticket 0023089 (#1062)
  Setup: Fixed #23080, #22723, #23079
  Setup: Fixed #23080, #22723, #23079
  Setup: Fixed #23080, #22723, #23079
  fixed mantis 0023023: ExtraTime is only available at limit of test passes is one ... description missing
  0023022: wrong byline on Manually selected Participants
  fixed mantis 0023017: Export of Test from workspace leads to an error
  fixLmMatchingPoints
  fixed 0022989: Export of Question Pool Test from workspace leads to an error
  0022972: CTM: not possible to delete own answers after all questions asnwered correctly
  ...

+++ Modules/Bibliographic/classes/class.ilObjBibliographic.php
@@ -115,7 +115,7 @@ class ilObjBibliographic extends ilObject2 {
 		global $DIC;
 
 		$upload = $DIC->upload();
-		if ($upload->hasUploads() && !$upload->hasBeenProcessed()) {
+		if ($_POST['override_entries'] && $upload->hasUploads() && !$upload->hasBeenProcessed()) {
 			$upload->process();
 			$this->deleteFile();
 			$this->moveUploadedFile($upload);
